Fix (recur) while capturing outer values
`(recur)` re-uses the closure from the original Arret fun. This is
normally a good idea except it doesn't update reg IDs from the values of
the original caller to the values inside the callee. This modifies
`load_from_closure_param` to update both the local values and the
closure. We then use that modified closure whenever we `(recur)`.