Don't lower and infer recursive (let)
Currently if a recursive `(let)` is used it will lower and type check
and then crash during when lowering to MIR. This is due to our MIR
fundamentally not supporting recursive values.

This follows what Clojure does and simply doesn't bind the new variable
inside its own definition. The only place this really makes sense is
mutually recursive functions which Clojure provides a `letfn` for. We
should support that eventually but not the general case of recursive
values.