Re-use test for phi-ing in branches
Right now for branches with constant booleans we end up returning
constant boxes and phi-ing them in to a `RegValue`. To use that
`RegValue` as a bool again we need to check its type tag. LLVM has
trouble optimising this away so it actually build type tag checks which
is quite wasteful. This happens with `(not)` which is implemented as an
`(if)`.

Instead if the results of our branches is either same as the test or the
reverse we can use the test's value. This is much easier to optimise and
crates the expected LLVM IR and assembler.