Don't forget regs containing boxed singletons
Right now `value_from_reg` and `value_with_arret_ty` with convert a
`RegValue` to a `Const` for the singleton true, false and nil boxes.
This was before we had `possible_type_tags` for a `RegValue` so this was
required to encode information about the type.

However, this generates clunky code for this typical Lisp list iteration
pattern:

```
(if (nil? input)
  input
  (recur ...)
```

Even though we know `input` is nil we actually reference the
`ARRET_NIL` global when returning `input`. This is correct but
inefficient.

Instead, only convert `RegValue` to a `Const` if we can't encode the
type information in the `RegValue`. This means we should be able to
reuse the reg directly instead of referencing its corresponding
singleton. Right now the only thing we can't properly encode is
`LitSym`.

This also makes a few places look at `possible_type_tags` for singleton
values. The most important one is `value_to_const` which was previously
depending on our const conversion.

This is nearly impossible to test with our fn ops framework because
it'll never pick a boxed singleton as a return type. We'd need to test
this indirectly by e.g. calling a function but that's fragile.