Fix selecting types based on union evidence
We would correctly handle selecting a type if the target was a union,
for example:

```
(U Int T) <= Float
```

Would correctly determine that the type of `T` is `Float`. However, this
didn't work properly in the other direction for complex types. For
example:

```
(List & T) <= (U (List Int Int) (List Int Int Int))
```

This should be a `(List & Int)` but we weren't looking inside the union
to find its component members. That means we couldn't find the
correspondence between the list structure on the left and the lists on
the right.

This is a trivial fix; we just need to iterate over the list members.
For scalar types this has the same result; we'll just end up
reconstructing the same union type. For complex types this allows us to
use the union members' structure.

The motivating breakage for this is function application. This used to
fail to compile because it couldn't find the type of `T`:

```
(concat & '((1 2 3) (4 5)))
```

If the lists on the right have the same types and length it would work
as they would collapse to the same non-union type. However, if the list
types aren't identical then selection would fail.