Add unchecked integer div & rem
If the denominator is constant we can rule out division by zero and
overflow except for `0` and `-1`.

Right now this generates identical LLVM IR but once we support checking
for overflow at runtime this can make us generate some more
straightforward LLVM IR. It's doubtful we'd catch anything LLVM wouldn't
so the generated assembler is likely to be identical as well.