Allow union number operands in intrinsics
Right now we'll only recognise an `Int` as an `Int` if that's its only
possible type tag. However, because it's being used as a number they
type checker has guaranteed it's a `Num`. This lets us use the more
permissive logic of "all things that can't be `Float` are `Int`" and
vice versa.