Allow discerned unions of records
Currently this turns in to `TopRecord`:

`(U Record1 Record2)`

This makes it very hard to do sum types as everything just turns in to
"unknown record".

This required beefing up of how we calculate `known_record_cons` in MIR.
The previous version wasn't quite accurate but mostly worked because
there could only be a single record in a union type. Once this
restriction was removed it became trivial to trick the MIR in to
thinking a union value had a known record type.