Generalise MIR known record cons to "type hints"
Right now if a `RegValue` can only have a single record cons we track
that inside the `RegValue.known_record_cons`. This is needed because the
record cons isn't encoded in our type tags so it must be tracked
separately. When this is set we can do a better job of optimising type
predicates and equality.

Now that we're gaining better support for vectors it would be really
useful to track known vector lengths as well. However, it's inefficient
and arduous to have a separate field for each possible type hint when
they're almost always mutually exclusive.

This instead extends to `TypeHint` that can either store a known record
cons or a known list length. The known list length is mostly used to
prove the concept; however, it does allow for some optimisations when
dealing with Arret tuple types like `(List Int Int Int)`.