Add codegen for loading inline vector members
While vectors will need to be replaced with persistent vectors, we'll
likely keep the same representation for small vectors (length of 0-3
inclusive). This means we can build codegen for loading their members
now with no wasted work.

This passes around a `known_vector_length` which isn't strictly needed
for codegen. However, persistent vectors will need this to know the
depth of the trie when loading a member. Add it in early and use it to
double check that the vector is inline.