Refactor PolymorphABI
Currently `PolymorphABI` is an `OpsABI` with some additional metadata.
This reflects how it evolved from a common usage pattern, but it's
actually quite awkward to deal with. This makes it its own first-class
type that can be converted to an `OpsABI` at the end.

This is in preparation for making functions polymorphic over function
values. However, this is a good cleanup on its own.