Support interning at runtime with global names
When interning a symbol at runtime (i.e. not in the JIT) we need to
ensure we don't duplicate an existing global name. Previously this was
unimplemented because there was no way to trigger interning at runtime.
However, `read-str` now allows arbitrary EDN to be read at runtime.

This is a quick hack. It adds a null global name at the end of the list
and then performs a linear search to find a matching name. However, it
will add the name to the local list of interned names so if the same
name is encountered again it won't need to be looked up until the next
GC.

A better solution would be modifying codegen to provide a more useful
data structure. Just doing a UTF-8 sort on the names would allow a
binary search. However, we need to be careful to not generate too much
data in the program to support this; most programs will never need this.