Emit global names in sorted order
This is a step towards improving our O(n) lookup of global names when
interning at runtime. Instead of assigning indexes to global names as
they're encountered this does a codegen analysis pass of all the used
global names first. They are then assigned indexes based on their
lexicographical order. The means the resulting global names array is
sorted when it's accessed at runtime.

Ideally this would allow us to do a binary search. However, we're still
missing an indication of the length of the global names array. Instead
this implements an early bailout if we "pass" our name when doing a
linear search. This halves the constant factor but is still a linear
search.