Implement binary search of global names
This takes advantage of the new sorted global names list to implement a
binary search when looking up a global name. This should significantly
speed up reading non-inline symbols from EDN at runtime when many
non-inline symbols were included in the binary.

This changes the global names representation slightly to start with a
length header instead of ending with a null entry. That should be
smaller by 8 bytes on 64bit platforms.