Refactor HIR imports
This splits the existing one-pass lowering in to three distinct passes:

1. Parsing consumes the import set and converts it to an AST

2. Loading loads the imported module

3. Filtering runs any filters (e.g. `only`) on the module's exports
   before inserting it in to the importing module's namespace.

The motivation is to allow modules with multiple imports to load their
dependencies in parallel. Currently they're loaded as they're
encountered which limits opportunities for parallelism. This refactor
doesn't run anything in parallel yet but should enable it later.

This also reduces the complexity of running the filters. This spends
some of the complexity budget by having the filters operate on
`Cow<Exports>`. This eliminates some useless clones that were happening
with the previous code.