Lower module imports specially
`(import)` is special because it can cause an Arret or RFI module to be
recursively lowered. Handling this inside `LoweringCtx` has been a total
mess and a blocker for supporting parallel/incremental HIR lowering.

Instead, this detects `import` early and enters a special path instead
of doing the normal module primitive logic. This isn't much of a win now
but is the beginning of dismantling `LoweringCtx` and parallel
compilation.