Hide uninferred HIR from compiler API
Right now we make our callers lower the entire program in to HIR and
then infer it as a second step. This makes it hard for us to run HIR
lowering and type inference in parallel.

Wrap the whole sequence as a `program_to_inferred_hir` and update our
callers. This is still done sequentially but gives us more freedom to
change behind the scenes.