Scope VarId to module
Currently type inference needs to run on an entire program at once in a
single thread. This is because all variables in the program are stored
in a huge `HashMap` inference can't be run concurrently without
prohibitively slow locking.

Theoretically when type inference runs on a module it should only be
inferring the variables for that modules, and it should only depend on
the completed inference from its imports. This means if we could
partition variables by module we could run a module's type inference
once we infer its imports.

This changes `VarId` to be represented as a `(ModuleId, LocalId)` tuple.
Every recursive definition context (either a module or REPL line) is
treated as a module. Once its inference is complete its inferred
variables are pushed on to a read only `HashMap` to be used in the
inference of modules that import it.

Right now this doesn't use the capability to actually infer modules in
parallel; this is just a refactor.