Add Promise and PromiseMap
Right now we can only load a single source file at once due to a mutex
around our `HashMap` of loaded files. We decided it was better to wait
for an existing load to complete instead of doing duplicate loads. This
isn't a major bottleneck at the moment but this limits us from
parallelising the compiler further with parallel module loads.

This introduces three new types:

- A `Promise` represents a computation that may be incomplete. Once its
  been completed it will continually clone its completed value.

- A `Complete` is a handle for completing promises

- A `PromiseMap` is a map of keys to `Promise`s. This is trickier to do
  than expected because we need to drop the lock on the map before we
  wait on the promise or else we're back to only supporting a single
  load at once.

This then uses a `PromiseMap<Path, CachedFile>` to allow us to load
different paths concurrently without loading the same file twice.

This same logic can be hopefully reused for implementing concurrent
compilation caches in other places in the compiler.