Prepare for incremental and parallel compilation
This removes the ad-hoc caching of our source files and RFI libraries
and replaces them with a centralised cache of compiled modules. This
provides the basis of incremental compilation by sharing the type
inferred modules between different compilations. At the moment this is
only used by our integration tests to avoid recompiling the `stdlib` for
each integration test. However, this will be essential for a good
Language Server experience.

This is also the basis of parallel compilation as we now try to collect
all of a module's imports and compile them in parallel. Previously we
would only import a single module at a time as they were encountered.
At the moment this doesn't work very well as we will block on the first
missing module which is going to be `[stdlib base]` by convention.
Follow up optimisations will attempt to improve parallelism.

The core data structure is a new `context::Module` which uniquely
identifies a module by name. `context.rs` now calls in to the lowering
and type inference code to drive a module's compilation up until type
inference. The stateful `LoweringCtx` and `InferCtx` have been removed
as now the state is centralised in `context.rs`. This makes parallelism
easier to reason about as now nearly the entire compiler is stateless
and thread safe except for `CompileCtx`, `EvalHirCtx` and codegen.

Right now this has little impact on the integration test's wall time but
reduces the overall user time. This indicates that we're doing a poor
job of extracting more parallelism but we're reducing duplicate work
that used to be spread across threads.

Before:
```
Benchmark #1: ../target/debug/deps/integration-76bf16ef3d4f5db4
  Time (mean ± σ):      6.382 s ±  0.123 s    [User: 7.158 s, System: 0.806 s]
  Range (min … max):    6.241 s …  6.550 s    10 runs
```

After:
```
Benchmark #1: ../target/debug/deps/integration-76bf16ef3d4f5db4
  Time (mean ± σ):      6.235 s ±  0.121 s    [User: 6.617 s, System: 0.831 s]
  Range (min … max):    5.978 s …  6.393 s    10 runs
```