Don't use Rustyline's ignore space
I thought this was ignoring lines comprising entirely of whitespace, not
lines starting with whitespace. This is a convention from Unix shells to
prevent sensitive lines from entered the history. It's unlikely someone
is doing something sensitive in the REPL and more likely they're miss
copying/pasting from somewhere. Switch it off to minimise annoyance.

Fortunately I never removed the "all whitespace" check when enabling
this so it doesn't need to be re-added.