Remove EMPTY_SPAN use from macro linker
`EMPTY_SPAN` is non-test code is a code smell because it will silently
propagate a useless `Span` as opposed to `Option<Span>` which requires
explicit handling.

One user of `EMPTY_SPAN` is the macro linker to special case finding
variables in the root of the pattern sequence. I initially used generics
to prove to the compiler we weren't using the `Span`. However, it's
trivial to just pass in the actual appropriate `Span` from the caller.