Introduce `span_for_rust_source_file`
This points to a zero length span with the specified filename. It's
intended for IR and code generated from the compiler that can't be
directly attributed to input Arret code. For example, the compiler will
generate synthetic functions on demand that can be shared between
multiple callsites.

This uses `span_for_rust_source_file` to replace a bunch of `EMPTY_SPAN`
instances. While this isn't as good as an actual source span, it should
at least point the developer towards which source file generated the
code.