Split up REPL driver
`subcommand/repl.rs` was a bit of a spaghetti mess. Unlike other
mega-files like `lowering.rs` and `eval_hir.rs` there is not a lot of
coupling between the different parts of this file. Split in to:

- `command` defines the builtin command strings and their parsing
- `history` finds and creates the history file
- `syntax` wraps `arret-syntax` with some higher level helpers
- `arret_helper` is an implementation of `rustyline::Helper`

The root module still contains the `interactive_loop` function that ties
the various pieces above together.