Don't linear search REPL completions
Whenever there's a new binding in the REPL we build a sorted
`Vec<DataStr>` with every bound name. This is then searched for
completions in two places:

1. Showing a grey hint after the cursor with the next completion

2. Cycling through completions when pressing tab

Right now we do a linear search over all bindings. Both searches involve
looking for a prefix with slight variations in logic afterwards

We can replace with with a binary search for start point + abort on
first mismatch. This takes us from O(log n) on the number of bound names
to O(log n + m) on the bound names and number of matches.