Enable incremental LSP document updates
This might be a slight performance win, but this is also building
machinery to allow us to produce diagnostics for syntax errors.

Using `codespan-diagnostics` would require registering each version of
the file which would rebuild line numbers. This code tries to be clever
and incrementally update the line numbers to track the changes. We can
use this to make reporting syntax errors low overhead while also kicking
off a full compilation concurrently.

We can also bypass doing syntax checks if we get an incremental update
after the previous syntax error. I'm unsure if this is worth the
complexity but it's an interesting opportunity.