Minimal LSP syntax diagnostics
This adds the concepts of document watchers that hook in to the document
lifecycle. On the various text synchronisation notifications we'll pass
an `Arc` of the current version of the document to the watcher. It's
then up to the watcher's implementation to handle scheduling,
concurrency, etc.

Right there's just one hardcoded syntax watcher that spawns a Tokio task
for each document. The task will run a syntax check on its document,
then dequeue the latest version and do it again. This avoids running
multiple syntax checks on the same document concurrently.

This doesn't have expected content logic that `arret-compiler` has so
the diagnostics are very basic at the moment.