Richer syntax error expected next description
Right now the compiler is doing a lot of heavy lifting for rendering
unexpected char/EOF errors based on `arret-syntax`'s error types. This
is unfortunate for the LSP as it's trying to bypass the compiler for the
hotpath of syntax checking.

This reshuffles the syntax error logic so the compiler is just exposing
the descriptions now provided by `arret-syntax`