Allocate an extra segment red zone cell
This is because variable sized boxes will have a Rust size of 32 bytes,
but we might allocate them only 16 bytes in the segment. If this happens
at the end of the segment we could create a pointer to a Rust object
extending past the allocation. This is UB and could generate code that
e.g. does a vector load into unallocated memory.