Get rid of (vector-push)/(vector-pop)
These were the first vector functions implemented due to the blog series
I was reading.

However, these are pretty much the worst way to use a vector. They
create a new vector every time which causes a lot of allocations and
garbage. Lists are a much better fit for a stack-like data structure.

`(vector-push)` has been replaced with a new `(vector-extend)` that
allows multiple values to be pushed at once. The existing
`(vector-take)` can be used take the initial part of the vector. Both of
these better fit with the performance characteristics of a persistent
vector as they encouraging working on multiple items at once.