Don't take refs to unreachable nodes
When using `(vector-first)` we try to share as much of the vector's
structure as possible. This means if there are intermediate nodes when
won't modify them to `null` out unreachable branches. This keeps extra
references around but avoids creating a new allocation.

When we push a new leaf, however, we already need to create a new node.
We can use this an opportunity to drop references to the unreachable
parts of the trie. This is also actually less work as we can just bail
out of initialisation early.