Implement bitwise operations
These match the name and semantics of the functions in Clojure exactly.

Clojure provides more functions (e.g. to set individual bits) but this
commit only adds the functions that correspond to an LLVM instruction.
The other functions should be able to be implemented in terms of these
and depend on LLVM to optimise them to platform-specific instructions
where available.

One exception to the LLVM-only rule is `(bit-not)` which is represented
in LLVM as a XOR against an all-ones value. This is a bit non-obvious so
even LLVM provides a helper builder function called `LLVMBuildNot`