Implement (repeat)
Unlike Clojure we don't support lazy or infinite collections. This means
we only accept the two argument form with an explicit length.

This is implemented in Rust so we can allocate the entire heap chunk
up-front. There is also a MIR eval version to generate a list of known
length and element value.