Remove Rayon usage
This used to be our go-to for parallelism. However, we shorty discovered
that any work that takes a lock will eventually deadlock on some
combination of hardware threads and input.

The only remaining users were in HIR lowering and RFI module loading.
When measuring elapsed time after removing Rayon:

1. HIR lowering is slightly faster on debug and 2x faster on release

2. RFI module loading loading is slightly slower on debug and slightly
   faster on release

Note that elapsed time doesn't take in to account CPU time wasted on
other threads; it seems Rayon was a loss even if time on other CPUs was
free. It also reduces our deps which should improve our own build time.