Allow intrinsics to simplify args
Currently intrinsics can either replace the entire function application
with a value or fall through the `stdlib`. This suits the majority of
intrinsics but there are a number of situations that would benefit from
just modifying the args passed to the `stdlib` function.

This replaces the builder's `Option<Value>` with a tri-state
`BuildOutcome` that allows for arg replacement. For now this is only
used by our numeric intrinsics.